Aug 7, 2024. Added algorithm for racing.
Todo
1. Turn logic when the next direction is different to current direction.
2. move2move
3. Save wall data logic

+++ 2809_hara/main/Rom.c
@@ -66,20 +66,20 @@ void ReadSensorData()
 
 	for(j = 0; j < SEN_NUM; j++)
 	{
-		g_s_sen[j].s_dist.s_formula.x0 = read_arr[i++] << 0;
-		g_s_sen[j].s_dist.s_formula.x0 |= read_arr[i++] << 8;
-		g_s_sen[j].s_dist.s_formula.x0 |= read_arr[i++] << 16;
-		g_s_sen[j].s_dist.s_formula.x0 |= read_arr[i++] << 24;
+		g_s_sen[j].s_dist.s_formula.x0 =  ((Uint32)read_arr[i++]) << 0;
+		g_s_sen[j].s_dist.s_formula.x0 |= ((Uint32)read_arr[i++]) << 8;
+		g_s_sen[j].s_dist.s_formula.x0 |= ((Uint32)read_arr[i++]) << 16;
+		g_s_sen[j].s_dist.s_formula.x0 |= ((Uint32)read_arr[i++]) << 24;
 
-		g_s_sen[j].s_dist.s_formula.x1 = read_arr[i++] << 0;
-		g_s_sen[j].s_dist.s_formula.x1 |= read_arr[i++] << 8;
-		g_s_sen[j].s_dist.s_formula.x1 |= read_arr[i++] << 16;
-		g_s_sen[j].s_dist.s_formula.x1 |= read_arr[i++] << 24;
+		g_s_sen[j].s_dist.s_formula.x1 =  ((Uint32)read_arr[i++]) << 0;
+		g_s_sen[j].s_dist.s_formula.x1 |= ((Uint32)read_arr[i++]) << 8;
+		g_s_sen[j].s_dist.s_formula.x1 |= ((Uint32)read_arr[i++]) << 16;
+		g_s_sen[j].s_dist.s_formula.x1 |= ((Uint32)read_arr[i++]) << 24;
 
-		g_s_sen[j].s_dist.s_formula.x2 = read_arr[i++] << 0;
-		g_s_sen[j].s_dist.s_formula.x2 |= read_arr[i++] << 8;
-		g_s_sen[j].s_dist.s_formula.x2 |= read_arr[i++] << 16;
-		g_s_sen[j].s_dist.s_formula.x2 |= read_arr[i++] << 24;
+		g_s_sen[j].s_dist.s_formula.x2 =  ((Uint32)read_arr[i++]) << 0;
+		g_s_sen[j].s_dist.s_formula.x2 |= ((Uint32)read_arr[i++]) << 8;
+		g_s_sen[j].s_dist.s_formula.x2 |= ((Uint32)read_arr[i++]) << 16;
+		g_s_sen[j].s_dist.s_formula.x2 |= ((Uint32)read_arr[i++]) << 24;
 	}
 }
